+++ src/main/kotlin/io/csbroker/apiserver/controller/v1/post/request/CommentCreateRequestDto.kt
@@ -1,5 +1,6 @@
 package io.csbroker.apiserver.controller.v1.post.request
 
 data class CommentCreateRequestDto(
+    val postId: Long,
     val content: String,
 )
